web_msn_revive/src/backend/requirements.txt
Flask>=3.1.2
Flask-SQLAlchemy>=3.1.1
Werkzeug>=3.1.5
Flask-Login>=0.6.3
pillow>=12.1.0
defusedxml>=0.7.1
gunicorn>=25.1.0